webnn/concat.html

<!doctype html>
<meta charset=utf-8>
<title>test concat operation</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-concat">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="https://webmachinelearning.github.io/webnn-polyfill/dist/webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  function testConcat(tensors, expected) {
    const inputs = [];
    const namedInputs = {};
    for (let i = 0; i < tensors.length; i++) {
      inputs.push(builder.input('input' + i, tensors[i].desc));
      namedInputs['input' + i] = new Float32Array(tensors[i].value);
    }
    const output = builder.concat(inputs, expected.axis);
    const graph = builder.build({output});
    const outputs = {
      'output': new Float32Array(utils.sizeOfShape(expected.shape)),
    };
    graph.compute(namedInputs, outputs);
    assert_array_approx_equals_ulp(outputs.output, expected.value, 0);
  }

  test(() => {
    const tensors = [
      {desc: {type: 'float32', dimensions: [2]}, value: [1, 2]},
      {desc: {type: 'float32', dimensions: [2]}, value: [3, 4]},
    ];
    const expected = {axis: 0, shape: [4], value: [1, 2, 3, 4]};
    testConcat(tensors, expected);
  }, 'concat 1d');

  test(() => {
    const tensors = [
      {desc: {type: 'float32', dimensions: [1, 2]}, value: [1, 2]},
      {desc: {type: 'float32', dimensions: [2, 2]}, value: [3, 4, 5, 6]},
    ];
    const expected = {axis: 0, shape: [3, 2], value: [1, 2, 3, 4, 5, 6]};
    testConcat(tensors, expected);
  }, 'concat 2d axis=0');

  test(() => {
    const tensors = [
      {desc: {type: 'float32', dimensions: [2, 1]}, value: [1, 2]},
      {desc: {type: 'float32', dimensions: [2, 2]}, value: [3, 4, 5, 6]},
    ];
    const expected = {axis: 1, shape: [2, 3], value: [1, 3, 4, 2, 5, 6]};
    testConcat(tensors, expected);
  }, 'concat 2d axis=1');

  test(() => {
    const tensors = [
      {desc: {type: 'float32', dimensions: [2, 2]}, value: [1, 2, 3, 4]},
      {desc: {type: 'float32', dimensions: [2, 2]}, value: [5, 6, 7, 8]},
    ];
    const expected = [
      {axis: 0, shape: [4, 2], value: [1, 2, 3, 4, 5, 6, 7, 8]},
      {axis: 1, shape: [2, 4], value: [1, 2, 5, 6, 3, 4, 7, 8]},
    ];
    for (const test of expected) {
      testConcat(tensors, test);
    }
  }, 'concat 2d');

  test(() => {
    const tensors = [
      {
        desc: {type: 'float32', dimensions: [2, 2, 2]},
        value: [1, 2, 3, 4, 5, 6, 7, 8],
      },
      {
        desc: {type: 'float32', dimensions: [2, 2, 2]},
        value: [9, 10, 11, 12, 13, 14, 15, 16],
      },
    ];
    const expected = [
      {
        axis: 0,
        shape: [4, 2, 2],
        value: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
      },
      {
        axis: 1,
        shape: [2, 4, 2],
        value: [1, 2, 3, 4, 9, 10, 11, 12, 5, 6, 7, 8, 13, 14, 15, 16],
      },
      {
        axis: 2,
        shape: [2, 2, 4],
        value: [1, 2, 9, 10, 3, 4, 11, 12, 5, 6, 13, 14, 7, 8, 15, 16],
      },
    ];
    for (const test of expected) {
      testConcat(tensors, test);
    }
  }, 'concat 3d');
</script>
